FIXED python script to parse CI_MESSAGE
Fixed the python script to parse the CI Message due to differences
between python3 (my dev env) and python2 (production env). Additionally,
adding a script, but failing to execute it in the pipeline does no good.
As such, I added sh steps to chmod+x the script and actually run it.

Finally, per @alivigni I uncommented the final sendMessage call in the
Jenkinsfile