ADD lib identifier to Jenkinsfile & PipelineUtils
ADDED library identifiers to Jenkinsfile-rf-pipeline and PipelineUtils
to allow for importing PipelineUtils and Utils libs respectively.

REMOVED default values for the following values in the Properties
section of Jenkinsfile-rf-pipeline, as they were also being set to the
same values in the setDefaultEnvVars() method.
 - HTTP_BASE
 - RSYNC_USER
 - RSYNC_SERVER
 - RSYNC_DIR
 - MAIN_TOPIC
 - FEDORA_PRINCIPAL

UPDATED setDefaultEnvVars() to evaluate env.MAIN_TOPIC to see if it
contains "staging". If so, we set env.RSYNC_DIR to
'fedora_atomic/staging', otherwise we set it to
'fedora_atomic/production' so that artifacts are rsynced to the proper
location.

UPDATED setDefaultEnvVars() to accept an envMap argument which allows
the passing of a map of key/value pairs to be passed which can overwrite
or add any additional arbitrary values.