UPDATED setting env.MAIN_TOPIC and env.RSYNC_DIR
Updated the logic in setDefaultEnvVars() to set env.MAIN_TOPIC and
env.RSYNC_DIR based on the presence and value of env.ghprbActualCommit.
If env.ghprbActualCommit is null or not equal to 'master' we set
env.MAIN_TOPIC and env.RSYNC_DIR to stage values respectively.

FIXED typo in library identifier call in Jenkinsfile-rf-pipeline.