FIXED body in podTemplate call in Jenkinsfile-rf-pipeline
Fixed the body call to podTemplate. Previously there were missing
opening and closing curly braces around the node call.

Updated the call to setMessageFields() in Jenkinsfile-rnester to avoid
issues with multiple assignment (see
https://issues.jenkins-ci.org/browse/JENKINS-38846 and
https://issues.jenkins-ci.org/browse/JENKINS-45575 for details of that
issue). setMessageFields now returns a map.

Updated the calls to sendMessage() in Jenkinsfile-rf-pipeline to use the
return of calling setMessageFields().

Fixed the assignment of env.RSYNC_DIR. Due to a typo, env.MAIN_TOPIC was
being overwritten the value of env.MAIN_TOPIC and env.RSYNC_DIR was not
being set in setDefaultEnvVars.