Add -R option to rsync for repo/
Added -R option to rsync when rsycing the repo dir. I think the rsync
which is occurring currently isn't using relative paths to copy the repo
path over, only its contents resulting in errors when referring to the
path.

Fixed typo in PipelineUtils for RSYNC_DIR if we're on master.