REMOVE -R argument to rsync
Removed the -R argument from rsync. Deteremined that the trailing slash
on the repo dir when rsyncing it over was resulting in a failure to
create the repo dir at the remote location. Upon removing the trailing
slash, -arv was sufficient to create the directory at the remote
location.

Updated the mkdir -p command to create the ${RSYNC_DIR}/repo empty
directory structure in one command, instead of two.