whitelist check for PR author or keyword commenter
Add two new options: --whitelist and --limit-whitelist

The first takes a file containing line-delimited github PR author names.
When specified, a PR will only be considered for triggering if the
author is in the referenced whitelist file.

The second option is related to --limit-keyword and requires --keyword.
When specified, it requires the keyword-commenter must be in the
whitelist file (the PR author is ignored).

Some other small changes were made to function-parameters to permit
simpler passing of options.  Also simplification of the triggering
conditional was made which will better support future addition of
conditions.

Signed-off-by: Chris Evich <cevich@redhat.com>