UPDATE Jenkinsfile to surround stages with withEnv
Updated the Jenkinsfile to surround stages with withEnv as a first step
to the eventual parallel run of some stages.

Updated the setStageEnvVars() method to return an ArrayList of strings
in the form "k=v" which can be used with the withEnv() call.
* Example:
* withEnv(pipelineUtils.setStageEnvVars(env.currentStage)){}

Updated the executeInContainer() method so that it will take an optional
argument (called vars), which is an ArrayList of "k=v" strings. Items in
this ArrayList will be added to the ArrayList returned from calling
env.getEnvironment(), already in the executeInContainer() method. This
avoids the need to set env vars repeatedly in the Jenkinsfile (such as
for rsync_paths, rsync_from, rsync_to)

Updated a few instances of variable interpolation to use more
"groovy-esque" variable interpolation in strings, instead of more
"java-esque" string concatenation.